+++ src/app/screens/settings/changePassword/index.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
+import { useSnackbar } from 'react-simple-snackbar';
+import styled, { useTheme } from 'styled-components';
 import TopRow from '@components/topRow';
 import BottomBar from '@components/tabBar';
 import { useTranslation } from 'react-i18next';
@@ -21,6 +22,7 @@ const PasswordContainer = styled.div((props) => ({
 
 function ChangePasswordScreen() {
   const { t } = useTranslation('translation');
+
   const [password, setPassword] = useState<string>('');
   const [confirmPassword, setConfirmPassword] = useState<string>('');
   const [error, setError] = useState<string>('');
@@ -29,7 +31,25 @@ function ChangePasswordScreen() {
   const { unlockWallet } = useWalletReducer();
   const navigate = useNavigate();
   const dispatch = useDispatch();
+  const theme = useTheme();
 
+  const options = {
+    position: 'bottom-right',
+    style: {
+      background: theme.colors.feedback.success,
+      borderRadius: theme.radius(2),
+      boxShadow: '0px 7px 16px -4px rgba(25, 25, 48, 0.25)',
+      color: theme.colors.background.elevation0,
+      fontStyle: theme.body_medium_m,
+      textAlign: 'center',
+    },
+    closeStyle: {
+      color: theme.colors.background.elevation0,
+      fontSize: '16px',
+    },
+  };
+
+  const [openSnackbar] = useSnackbar(options);
   const handleBackButtonClick = () => {
     navigate('/settings');
   };
@@ -54,11 +74,11 @@ function ChangePasswordScreen() {
       setError('');
       const encryptedSeed = await encryptSeedPhrase(seedPhrase, password);
       dispatch(storeEncryptedSeedAction(encryptedSeed));
-      navigate('/settings', {
-        state: {
-          passwordUpdated: true,
-        },
-      });
+      openSnackbar(
+        t('SETTING_SCREEN.UPDATE_PASSWORD_SUCCESS'),
+        10000,
+      );
+      navigate('/settings');
     } else {
       setError(t('CREATE_PASSWORD_SCREEN.CONFIRM_PASSWORD_MATCH_ERROR'));
     }
