+++ src/app/screens/settings/index.tsx
@@ -4,15 +4,14 @@ import useWalletSelector from '@hooks/useWalletSelector';
 import XverseLogo from '@assets/img/settings/logo.svg';
 import ArrowIcon from '@assets/img/settings/arrow.svg';
 import ArrowSquareOut from '@assets/img/arrow_square_out.svg';
-import CheckSquare from '@assets/img/settings/check_square.svg';
 import BottomBar from '@components/tabBar';
 import { PRIVACY_POLICY_LINK, TERMS_LINK, SUPPORT_LINK } from '@utils/constants';
-import { useLocation, useNavigate } from 'react-router-dom';
-import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useState } from 'react';
 import PasswordInput from '@components/passwordInput';
 import useWalletReducer from '@hooks/useWalletReducer';
 import SettingComponent from './settingComponent';
-import ResetWalletPrompt from './resetWallet/resetWalletPrompt';
+import ResetWalletPrompt from './resetWallet';
 
 declare const VERSION: string;
 
@@ -26,34 +25,6 @@ const Container = styled.div`
     display: none;
   }
 `;
-
-const PasswordChangedSuccessContainer = styled.div((props) => ({
-  display: 'flex',
-  flexDirection: 'row',
-  alignItems: 'center',
-  justifyContent: 'space-between',
-  paddingLeft: props.theme.spacing(10),
-  paddingRight: props.theme.spacing(10),
-  paddingTop: props.theme.spacing(5),
-  paddingBottom: props.theme.spacing(5),
-  marginTop: props.theme.spacing(5),
-  marginLeft: props.theme.spacing(13),
-  marginRight: props.theme.spacing(13),
-  marginBottom: props.theme.spacing(5),
-  background: props.theme.colors.feedback.success,
-  borderRadius: props.theme.radius(2),
-  boxShadow: '0px 7px 16px -4px rgba(25, 25, 48, 0.25)',
-}));
-
-const PasswordChangedSuccessText = styled.h1((props) => ({
-  ...props.theme.body_medium_m,
-  textAlign: 'center',
-  color: props.theme.colors.background.elevation0,
-}));
-
-const Button = styled.button({
-  background: 'none',
-});
 
 const ResetWalletContainer = styled.div((props) => ({
   display: 'flex',
@@ -71,8 +42,6 @@ const LogoContainer = styled.div((props) => ({
 
 function Setting() {
   const { t } = useTranslation('translation', { keyPrefix: 'SETTING_SCREEN' });
-  const location = useLocation();
-  const [passwordChangeSuccessMessage, setPasswordChangeSuccessMessage] = useState<boolean>(false);
   const [showResetWalletPrompt, setShowResetWalletPrompt] = useState<boolean>(false);
   const [showResetWalletDisplay, setShowResetWalletDisplay] = useState<boolean>(false);
   const [password, setPassword] = useState<string>('');
@@ -80,12 +49,6 @@ function Setting() {
   const { fiatCurrency, network } = useWalletSelector();
   const navigate = useNavigate();
   const { unlockWallet, resetWallet } = useWalletReducer();
-
-  useEffect(() => {
-    if (location?.state?.passwordUpdated) {
-      setPasswordChangeSuccessMessage(true);
-    }
-  }, [location]);
 
   const openTermsOfService = () => {
     window.open(TERMS_LINK);
@@ -113,21 +76,15 @@ function Setting() {
 
   const openUpdatePasswordScreen = () => {
     navigate('/change-password');
-    setPasswordChangeSuccessMessage(true);
   };
 
   const openResetWalletScreen = () => {
     setShowResetWalletPrompt(false);
     setShowResetWalletDisplay(true);
-    // navigate('/reset-wallet');
   };
 
   const openResetWalletPrompt = () => {
     setShowResetWalletPrompt(true);
-  };
-
-  const openDismissSuccessMessage = () => {
-    setPasswordChangeSuccessMessage(false);
   };
 
   const onResetWalletPromptClose = () => {
@@ -153,16 +110,6 @@ function Setting() {
       setError(t('INCORRECT_PASSWORD_ERROR'));
     }
   };
-
-  const showSuccessMessage = (
-    <PasswordChangedSuccessContainer>
-      <img src={CheckSquare} alt="xverse logo" />
-      <PasswordChangedSuccessText>{t('UPDATE_PASSWORD_SUCCESS')}</PasswordChangedSuccessText>
-      <Button onClick={openDismissSuccessMessage}>
-        <PasswordChangedSuccessText>{t('OK')}</PasswordChangedSuccessText>
-      </Button>
-    </PasswordChangedSuccessContainer>
-  );
 
   return (
 
@@ -237,7 +184,6 @@ function Setting() {
           <SettingComponent text={t('VERSION')} textDetail={VERSION} />
           <ResetWalletPrompt showResetWalletPrompt={showResetWalletPrompt} onResetWalletPromptClose={onResetWalletPromptClose} openResetWalletScreen={openResetWalletScreen} />
         </Container>
-        { passwordChangeSuccessMessage && showSuccessMessage}
         <BottomBar tab="settings" />
       </>
     )
